requirements.txt
requests
langgraph 
langsmith
langchain_anthropic
python-dotenv
ipython